Rename Speaker encoder module to encoder
